requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
chromadb==0.4.15
langchain==0.0.350
langchain-community==0.0.10
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
aiofiles==0.24.0
jinja2==3.1.2 